bvh: Make nodes traversal closer to linear access
Nodes now are put into an array, sorted by in-order traversal. Each node
visit (left / right) cuts each partition in half, so it just evaluates
the whole partition as an array once it doesn't have any more AABBs to
check.

This implementation doesn't worry about SAH costs yet, so it makes a
full tree for everything. I'll redo the SAH costs later to avoid
partitioning too much.

Now visiting takes less cache misses on the deeper levels, since the
deeper levels are kept local between child-parent-child jumps, finishing
left and jumping to right.

Why in-order?
The visiting is pre-order (first parent, then all children).

What the processor sees is that we make a couple of accesses (box, left,
right), we go left immediately (as if iterating an array backwards),
then we index another place that it doesn't have in cache (objects, when
root == -1), then we make another jump (right), before going again left.

It stands to reason then that the in-order arrangement will put all
nodes that are to the left of some higher node to the left of it, which
means that the prefetcher sees a somewhat constant stride, while the
go-back-and-do-right jumps still hit near addresses on the depths of the
traversal, which is where most of the nodes will be.

If I'd sort the nodes according to pre-order traversal, then the deeper
the traversal goes, the further apart groups of children from the same
node are, which means that the processor misses on each recursion.