bvh: Return a collection instead of a 'single' object
Make clear to the user code that it's getting a collection,
so that it can decide how to wrap it. This will help to keep simplifying
the collection interface so that it's more coupled with the renderer.

The next thing to do is to make 'hittable_list' just a builder, and
decide how we want to get the collection from it. When everyone uses it
to build the geometry layout, then I can start swapping bits to do
things like separate materials, or joining everything into a single
collection with a layering scheme.