noise: Simplify & reduce memory footprint
Removed the clamping in perlin::turb so that the big clamped hole
doesn't exist anymore and we get a more sine-like form. Still need to
play with the parameters to get that marble-like appereance.

Since each noise sample only occupies a grayscale channel, it made sense
to make another view just for the grayscale and multiply it by
broadcasting.