Make collection & hittable concept-based
This uses the dyn-trait pattern for collections & hittables, which lets
me relax the type requirements in BVH construction, since it only needs
to know about bounding boxes.

Makes render() a template, but since it's only used in one instance the
compiler is smart enough to not re-instance it for all the types....
hopefully.

Thanks to this I control _exactly_ where vtables are introduced instead
of helplessly adding them in every object.