Decouple BVH from type of object
BVH 'propagation' is now modelled as a span producer with internal state
for the now unavailable 'recursion' stack.

The view-specific propagation is modelled by a different struct
'over_hittables' so BVH can stay the same for all types.

This will make it easier to  make it wrap a group of geometries with a
single texture/material, speeding up the process out of removing memory.