Move scene to compile-time switch approach
This way the compiler can figure out how to optimize for the different
cases and the types are implicit, hence no runtime switches for
collections/hittables. The `tuple_wrapper` struct allows me to have a
heterogeneous list of collections/singles and not have to worry about
wrapping them. Should do the same in the rest of scenes!

Also, the rendering bugs from the last commits are fixed, thanks to
removing yet more things from interfaces