bvh: Move static buffers dependency to bvh::over
Next step with this is to make bvh::over accept them as dependencies,
although that may take more concept work (or not!). The idea is that the
state buffer (read/write) is private but the nodes are shared (read-only).

These steps aim to make it easier to move into a queue based behavior
where everything owned by a single thread is explicit in the render()
method and thus a compute shader is easier to write.