Separate objects from collections
Objects return whether they are hit, collections select what object has
been hit.

1. Makes BVH not a 'hittable' so that BVHs have to be processed
   explicitly by 'hittable_list'.
2. 'hittable' now has a guaranteed material in it. Now transforms are
    not ideal 'hittable's but they also don't fall into the interface
    for selectors.
3. The renderer expects a collection, which in some time should only be
   a `hittable_list`.
4. `hittable_list` exports an interface to yield spans so that objects
   created can be modified in-place before the next modification to the
   list's size.

Performance is slightly worse due to extra indirections for all the
transform modifiers that now must apply to individual objects. Note
that, if transforming and applying the inverse results in duplicate
expensive work, a layering system with transform ranges can be done once
all the objects are in a single array. The objects could also be
scattered across a typed set of arrays, in which case the transform
ranges would also be scattered across the same types.