Calculate UV coordinates at the end of the intersection
They aren't needed during the intersection process. We gain a 23% (3
seconds less) improvement from previous commit time, accounting for a
total of 54% since the parallel version of the original.

The sphere timing dilemma is still not solved, even though there are
much less calls to the costly atan2. The timings are still spread
between 10s of ns to ~3µs.