Join transforms
This removes two layers of indirection (recursion, pointer-to-pointer) through geometries.
It also forces the order of transforms, since main.cc was using the same order everywhere.
Saves another ~500ms.

There is still space for improvement. The de-transform is not needed
till the end and takes ~80% of the transformed hit time. Could be huge
to reduce drastically the frequency of calls by only sampling it when
required.