Merge all BVH trees
The node insertion order is key for this: It guarantees that
the next valid memory location in the nodes array is either the left
node of the current node or the right subtree of the next toplevel node.

That also means that once we finish traversing a tree we automatically
traverse the next!