use pass-by-value in getUVs
Also requires the normal since renderer.cc may change the normal based
on the face information.

Saves 4% but also gives much better hints on how things are passed
around. I may do this with more things as long as the depth is not too
big.

Kind of gives away that transform does not really fit in geometry and
that maybe it's time to support some way of either always including a
transform set (given that identity transform exist) and/or separating the
geometry information into ones that require such transforms (or some
number of it) and ones that require other counts of transforms, if
computing the transforms affects performance (i.e checking if list ptr
is null)