Transpose geometrySim to work on collections of rays
This is the initial step in a series of changes to convert the full
simulation to work in batches of rays. The reasoning behind this is
that the target volume looks at 10k samples per pixel, so being
able to bulk process good loads of rays faster will benefit us a lot.

get_ray() is now called in a tight loop (transform), so it may as well
be eligible for bulk-style optimization.

The change now calls hittable_list::sampleConstantMediums and
hittable_list::hitSelect in tight loops, which means I can start doing
more of the same things: abstracting each end, transposing and inlining.

Since I'm starting to put things in tight loops, benchmarks other than
producing the image could come in clutch. Specially to avoid 20%
regressions, Gbench and/or Catch2 benchmarks could help me avoid
significant perf downgrades, and it wouldn't be bad to include
e2e tests for bugs I end up finding in the final image. With all the
noise it may be hard, but I could also explore some image filters
that can make testing for features easier.

The attenuation queues are filled in "parallel", which means I could
probably make each att-sample have an ID (pixel, sample no) so that I
can join them later.  The current implementation will basically scan the
queues and copy over the elements to defragment the matrix into a bunch
of linear arrays. From there the sampling is the same, done in
bulk-style. The ID idea would include a "termination" att-sample, where
I know there are no more color samples for said ray and I can move the
ray to yet another scan queue where the colors can be aggregated and
pushed into the final image.