Convert sampleConstantMediums to bulk processing rays
This exposes geom.traverse() as a big candidate for converting to
bulk-processing, and one of the tasks for it is going to be partitioning
depending on the kind, which can be then passed to be done before
sampling.

Another interesting candidate is random_double(), because we make up to
SPP samples (10k in the big case) per medium simulation.

It would be interesting to first hit all the objects and make the
comparison scan later.

The compiler is no longer optimizing for non-null pointer
because we're doing a block-wise separation, so the flow of
if (ptr) is indirect through a (series of) partition(s).

The algorithm is simpler here as it is going to act on all rays anyway.
We hit, compare with last and update the best one.