Convert BVH hit to bulk processing
BVH hit now keeps the node constant so we can traverse all
relevant rays in a single go.

It's quite slower than the previous version, partially because there is
a branch for leaf nodes. This branch can be also removed from the above
loop, by joining all the leaf nodes at the end of the nodes array. The
only problem right now is that node_ends[leaf node index] is the start
of the next tree.

So I should just partition each 'tree' section, such that the node
indices for the roots of each tree are kept intact.