aabb: Use avx2 with head-block-tail in getNormals
Converted the routine to work with 8-wide blocks. Adds a utility header
for SIMD operations like rotations and finding alignment offsets.

Since the routine (and most that use vectors) acts on 3-wide blocks
(vec3), in order to work with 8-wide blocks I have to first try to
downgrade and use float by float, and if something that relates to the
axes is used, then I have to use the MCM (3 * 8 = 24) as base block if
I don't want anything to change.

After the code is converted to use 8-wide blocks in SIMD fashion
(without SIMD registers), the compiler is smart enough to vectorize
the loops, which makes swapping the code for SIMD have very low gains
over the compiler-managed code.

This whole process took around ~15% off the original code. Again, most
of it was achieved by converting the routine first to float-by-float and
then using alignment and strides of 8 in loops.