sphere: Remove unneeded checks.
Rays are now always normalized, this way an interesting chunk of compute gets removed.

Discriminant from quadratic root finding doesn't need to be checked,
since it's already indirectly checked in hitSpan

We always want the side of the sphere that is closest to the ray. The
one that is nearest will just make the ray pass through. That was a
problem with 10 rays per pixel but now that I've moved SPP to >100 it's
completely covered up by the rest. Speeds up by at least 10%.